refactor to use metrical time
This is a major revamp to change the time used in
the looper and layer to be metrical time, and then
scale that time via a tempo to wall clock time.

As an aside, the layer timing analysis routines were
split out to analysis.cpp, and some of them are currently
disabled.